support/kernel-config: enable zswap
From https://wiki.archlinux.org/title/Zswap:
"[zswap](https://en.wikipedia.org/wiki/zswap) is a kernel feature that provides a compressed RAM cache for swap pages. Pages which would otherwise be swapped out to disk are instead compressed and stored into a memory pool in RAM. Once the pool is full or the RAM is exhausted, the least recently used ([LRU](https://en.wikipedia.org/wiki/Cache_replacement_policies#Least_recently_used_.28LRU.29)) page is decompressed and written to disk, as if it had not been intercepted. After the page has been decompressed into the swap cache, the compressed version in the pool can be freed."
"The difference compared to [zram](https://wiki.archlinux.org/title/Zram) is that *zswap* works in conjunction with a [swap](https://wiki.archlinux.org/title/Swap) device while *zram* with created swap on top of it is a swap device in RAM that does not require a backing swap device."