overlay: Move mobile-nixos.stage-1 into another overlay...
This seems like a lot, but this is going to simplify some stuff we need
to do regarding conditionally overriding inputs.

Previously, we were always overriding libxkbommon and libinput. It was
fine for stage-2 usage too, until it was not anymore.

Next, we moved to `replaceDependencies`, which at the time was not
noticed to cause IFD. This is problematic because IFD causes spurious
rebuilds that may not be possible depending on some contexts. It also
will block evaluation needlessly. On top of that, `replaceDependencies`
would not error out on dependencies that were not matched. So that's a
no-go imo.

The "next best thing" is to meet-up halfway.

We make this new overlay that allows *anything* to be overridden as
needed in the context of `callPackage`, using the usual cascade of
overlays. And then we hoist only the `mobile-nixos.stage-1` attribute
set that were part of that overlay.

The only less-than-magical part we're left with is dealing with two
discrete `mobile-nixos.script-loader` attributes. We *could* probably
rethink how the consumers of `mobile-nixos.script-loader` are made so
they could naturally inherit from the right context, but that's
something for a major refactor, and not a fix like here.