modules/kernel-config: disable legacy feature on new kernels
`CONFIG_IP_NF_RAW` depends on `CONFIG_IP_NF_IPTABLES_LEGACY`, which depends on `CONFIG_NETFILTER_XTABLES_LEGACY`, which is off by default.
The help text of `CONFIG_NETFILTER_XTABLES_LEGACY` is: "Say Y here if you still require support for legacy tables. This is required by the legacy tools (iptables-legacy) and is not needed if you use iptables over nftables (iptables-nft). Legacy support is not limited to IP, it also includes EBTABLES and ARPTABLES."
`CONFIG_IP6_NF_RAW` depends on `CONFIG_IP6_NF_IPTABLES_LEGACY`, which also depends on `CONFIG_NETFILTER_XTABLES_LEGACY`.
Instead of turning on `CONFIG_NETFILTER_XTABLES_LEGACY`, let's keep this off on newer kernels for now. If someone needs it, we can always turn it back on. There is currently no device with a kernel this new in Mobile NixOS but the next commit will update the PinePhone kernel to 6.17.